+++ climpred/classes.py
@@ -2825,6 +2825,7 @@ class HindcastEnsemble(PredictionEnsemble):
                 )
             metric_kwargs["cv"] = cv
 
+        old_coords = self.get_initialized().coords
         self = func(
             self,
             alignment=alignment,
@@ -2849,4 +2850,22 @@ class HindcastEnsemble(PredictionEnsemble):
                 self._datasets["initialized"] = self._datasets["initialized"].dropna(
                     "init"
                 )
+        # avoid single-item lead dimension being dropped, see #771
+        for c in ["lead", "init"]:
+            if c not in self.get_initialized().dims and c in old_coords:
+                if old_coords[c].size == 1:
+                    self._datasets["initialized"] = self._datasets[
+                        "initialized"
+                    ].assign_coords({c: old_coords[c]})
+        if "valid_time" in self.get_initialized().coords:
+            if len(self.get_initialized().coords["valid_time"].dims) == 1 and set(
+                ["init", "lead"]
+            ).issubset(set(self.get_initialized().dims)):
+                self._datasets["initialized"].coords[
+                    "valid_time"
+                ] = add_time_from_init_lead(
+                    self.get_initialized().drop("valid_time")
+                ).coords[
+                    "valid_time"
+                ]
         return self